Updating foxhound version

+++ dom/workers/sharedworkers/SharedWorkerManager.cpp
@@ -16,8 +16,7 @@
 #include "nsIPrincipal.h"
 #include "nsProxyRelease.h"
 
-namespace mozilla {
-namespace dom {
+namespace mozilla::dom {
 
 // static
 already_AddRefed<SharedWorkerManagerHolder> SharedWorkerManager::Create(
@@ -63,6 +62,12 @@ bool SharedWorkerManager::MaybeCreateRemoteWorker(
     const RemoteWorkerData& aData, uint64_t aWindowID,
     UniqueMessagePortId& aPortIdentifier, base::ProcessId aProcessId) {
   ::mozilla::ipc::AssertIsOnBackgroundThread();
+
+  // Creating remote workers may result in creating new processes, but during
+  // parent shutdown that would add just noise, so better bail out.
+  if (AppShutdown::IsInOrBeyond(ShutdownPhase::AppShutdownConfirmed)) {
+    return false;
+  }
 
   if (!mRemoteWorkerController) {
     mRemoteWorkerController =
@@ -119,6 +124,10 @@ void SharedWorkerManager::AddActor(SharedWorkerParent* aParent) {
     Unused << aParent->SendNotifyLock(true);
   }
 
+  if (mWebTransportCount) {
+    Unused << aParent->SendNotifyWebTransport(true);
+  }
+
   // NB: We don't update our Suspended/Frozen state here, yet. The aParent is
   // responsible for doing so from SharedWorkerParent::ManagerCreated.
   // XXX But we could avoid iterating all of our actors because if aParent is
@@ -149,6 +158,12 @@ void SharedWorkerManager::Terminate() {
   ::mozilla::ipc::AssertIsOnBackgroundThread();
   MOZ_ASSERT(mActors.IsEmpty());
   MOZ_ASSERT(mHolders.IsEmpty());
+
+  // mRemoteWorkerController creation can fail. If the creation fails
+  // mRemoteWorkerController is nullptr and we should stop termination here.
+  if (!mRemoteWorkerController) {
+    return;
+  }
 
   mRemoteWorkerController->Terminate();
   mRemoteWorkerController = nullptr;
@@ -249,6 +264,22 @@ void SharedWorkerManager::LockNotified(bool aCreated) {
   }
 };
 
+void SharedWorkerManager::WebTransportNotified(bool aCreated) {
+  ::mozilla::ipc::AssertIsOnBackgroundThread();
+  MOZ_ASSERT_IF(!aCreated, mWebTransportCount > 0);
+
+  mWebTransportCount += aCreated ? 1 : -1;
+
+  // Notify only when we either:
+  // 1. Got a first WebTransport
+  // 2. The last WebTransport goes away
+  if ((aCreated && mWebTransportCount == 1) || mWebTransportCount == 0) {
+    for (SharedWorkerParent* actor : mActors) {
+      Unused << actor->SendNotifyWebTransport(aCreated);
+    }
+  }
+};
+
 void SharedWorkerManager::Terminated() {
   ::mozilla::ipc::AssertIsOnBackgroundThread();
 
@@ -314,5 +345,4 @@ SharedWorkerManagerWrapper::~SharedWorkerManagerWrapper() {
                          mHolder.forget());
 }
 
-}  // namespace dom
-}  // namespace mozilla
+}  // namespace mozilla::dom
